// src/game.services/store.services/store.interface.ts
export interface StoreCardData {
    name: string;
    description: string;
    group: string;
    era: string;
    scoreBoost: string;
    healBoost: string;
    skill: string;
    stars: string;
    slot: string;
    tier: string;
    boostCount: string;
    awakenCount: string;
    breakthrough: boolean;
    artist: string; 
    position: string;
    position2: string;
    rarity: string;
    level: string;
    experience: string;
    type: "ERC1155" | "ERC721" | "metaplex";
    image: string;
    imageByte: string;
    uri: string;
    owner: string;
    tokenId: string
    id: string;
    listingId: number | string;
    cardAddress: string;
    uploader: string;
    supply: string;
    quantity: 1
    quantityOwned: string
    startTime: string;
    endTime?: string;
    lister: string;
    pricePerToken: number
    currencyName: string
}


/**
 * Interface for buying card data.
 * @interface BuyCardData
 * @property {string} uri - URI of the card.
 * @property {number} listingId - Listing ID of the card.
 * @property {string} price - Price of the card.
 */
export interface BuyCardData {
    uri: string;
    listingId: number;
    price: string;
}




/**
 * Interface for buying card upgrade data.
 * @interface BuyCardUpgradeData
 * @property {string} uri - URI of the upgrade.
 * @property {number} listingId - Listing ID of the upgrade.
 * @property {string} quantity - Quantity of the upgrade.
 */
export interface BuyCardUpgradeData {
    uri: string;
    listingId: string;
    quantity: string
    price: string
}
// ...
export interface StoreCardUpgradeData {
    currencyName: string;
    editionAddress: string;
    endTime: string;
    experience: number;
    id: string;
    image: string;
    lister: string;
    listingId: number;
    minted: boolean;
    owner: string;
    pricePerToken: number;
    quantity: number;
    quantityOwned: string;
    startTime: string;
    supply: string;
    tier: string;
    tokenId: string;
    type: string;
    uploader: string;
    uploaderBeats: string;
    uri: string;
}


export interface CardUpgradeAsset {
	id: string;
	uri: string;
	name: string;
	image: string;
	experience: number;
	itemType: string;
	minted: boolean;
	quantity: number;
	tier: string;
	uploader: string;
}




export interface StorePackData {
    currencyName: string;
    description: string;
    endTime: string; // Consider using a Date type if you need date manipulation
    id: string;
    image: string;
    lister: string;
    listingId: number;
    name: string;
    pricePerToken: number;
    quantity: number;
    startTime: string; // Consider using a Date type if you need date manipulation
    tokenId: string;
    type: string;
    uploader: string;
    uri: string;
}