// src/user.services/user.service.interface.ts
export interface PlayerStats {
    level: number;
    playerExp: number;
    availStatPoints: number;
    rank: string;
    statPointsSaved: {
      mainVocalist: number;
      rapper: number;
      leadDancer: number;
      leadVocalist: number;
      mainDancer: number;
      visual: number;
      leadRapper: number;
    };
  }
// ...
export interface UserData {
    properties: {
      signupDate: number
      accountType: "beats",
      userId: string;
      username: string;
      password: string;
      localWallet: string;
      localWalletKey: string;
      playerStats: PlayerStats;
      suspended: Suspended
      country: string;
      deviceId: string;
      inventorySize: any;
      smartWalletAddress: string;
      encryptedAccessKey: string;
      preferredServer: string;
    };
  }
// ...
export interface Suspended {
  until: number | null
  reason: string 
}